Final copy hw2 and hw2_test file
The assignment took about an hour to finish up, with minimum debugging.
The readings didn't take long, maybe one evening. The ICUPIP book is a
much tougher read. The lecture did not quite get to FOR loops, so we had
to use ICUPIP, but the assignment did build off itself.